requirements.txt
streamlit
pandas
numpy
Pillow
scikit-learn
tensorflow
